+++ src/UI/WB.UI.Designer/questionnaire-app/tools/vite-plugin-localization.js
@@ -4,8 +4,9 @@ export default (userOptions = {}) => {
     return {
         name: "vite-plagin-localization",
         enforce: "pre",
+        userOptions: userOptions,
         buildStart(options){
-            this.localization = new LocalizationBuilder(options);
+            this.localization = new LocalizationBuilder(userOptions);
             this.localization.prepareLocalizationFiles();
         }
     }
